Move models to core
