+++ src/core/device/mod.rs
@@ -1,0 +1,14 @@
+mod archiver;
+mod backup_requirement;
+mod device_factories_registry;
+mod question;
+mod secondary_device;
+
+pub use archiver::{ArchiveError, ArchiveWriter};
+pub use backup_requirement::{BackupRequirementClass, SecurityLevel};
+pub use device_factories_registry::DeviceFactoryRegistry;
+pub use question::{Question, QuestionType};
+pub use secondary_device::{Device, DeviceFactory, DeviceFactoryKey};
+
+#[cfg(test)]
+pub use secondary_device::{MockDevice, MockDeviceFactory};
