+++ src/core/project/mod.rs
@@ -1,0 +1,5 @@
+mod project;
+mod project_status;
+mod projects_scan;
+
+pub use project::{Project, ProjectTrackingStatus};
